additional updates to user controller

+++ src/controllers/userController.ts
@@ -1,4 +1,5 @@
 import User from '../models/user.js';
+import Thought from '../models/thought.js';
 import { Request, Response } from 'express';
 
   export const getUsers = async(_req: Request, res: Response) => {
@@ -54,17 +55,56 @@ import { Request, Response } from 'express';
 
   export const deleteUser = async (req: Request, res: Response) => {
     try {
-      const user = await User.findOneAndDelete({ _id: req.params.userId });
+      const user = await User.findbyIdAndDelete({ _id: req.params.userId });
 
       if (!user) {
         return res.status(404).json({ message: 'No user with that ID' });
       }
-
+      
+      if (user.thoughts.length >0)
       await Thought.deleteMany({ _id: { $in: user.thoughts } });
       res.json({ message: 'User and associated apps deleted!' })
     
     } catch (err) {
       res.status(500).json(err);
       
+    }
+  }
+
+
+  export const addFriend = async (req: Request, res: Response) => {
+    try {
+      const user = await User.findByIdAndUpdate(
+        req.params.userId,
+        { $addToSet: { friends: req.params.friendId } }, // Use $addToSet to avoid duplicates
+        { new: true }
+      );
+  
+      if (!user) {
+        return res.status(404).json({ message: 'No user with that ID' });
+      }
+  
+      res.json(user);
+    } catch (err) {
+      res.status(500).json(err);
+    }
+
+}
+
+export const removeFriend = async (req: Request, res: Response) => {
+    try {
+      const user = await User.findByIdAndUpdate(
+        req.params.userId,
+        { $pull: { friends: req.params.friendId } }, // Use $pull to remove the friend
+        { new: true }
+      );
+  
+      if (!user) {
+        return res.status(404).json({ message: 'No user with that ID' });
+      }
+  
+      res.json(user);
+    } catch (err) {
+      res.status(500).json(err);
     }
   }
